services/classhub/requirements.txt
Django==5.0.8
psycopg[binary]==3.2.1
django-environ==0.11.2
whitenoise==6.7.0
